backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Telegram
aiogram==3.17.0

# AI
openai==1.59.5
httpx==0.28.1

# Cache & Tasks
redis==5.2.1
celery==5.4.0

# Utils
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Supabase
supabase==2.11.0
